refactor(experimental): a low-level series of helpers to confirm transactions (#1686)
# Summary

There are some transaction confirmation strategies on which there is some disagreement, like whether to poll the server for a signature's status or to subscribe for changes to it.

In this PR we commit the _uncontroversial_ parts of the strategy; that you will have to wait for the signature to be confirmed (somehow) and fail if the block height exceeds the `lastValidBlockHeight`.

These functions leave the implementation of _how_ to do that up to the caller. This will enable us to rebuild the default strategy on top of this, but will also allow people to supply their _own_ strategies, as appropriate for their use case.

# Test Plan

```
cd packages/library
pnpm test:unit:browser
pnpm test:unit:node
```