refactor(experimental): rename signature confirmation strategy using the word ‘recent’ (#1744)
# Summary

A common pitfall for folks is trying to use `confirmTransaction()` to ‘confirm’ really old transactions that have fallen out of the status cache. You should not do this; you should use `getSignatureStatuses([...], {searchTransactionHistory: true})` for that.

In this PR, we rename _all_ of this stuff using the word ‘recent’ to drive that home.