refactor(experimental): modify account resolver to not send a request for only an address
This PR adds an interceptor on the actual query passed into the resolver. The interceptor is designed to read which fields have been requested for an account, and determine the necessary action.

If _only_ the address has been requested, it's already in the query - either inside a provided query variable or in a previously obtained return payload. So, there's no need to make an RPC call to `getAccountInfo` if the user only needs the address.

If any other fields are requested, with or without `address`, the resolver will go to the RPC.