refactor(experimental): align program account types with account types
This PR decouples RPC-GraphQL's response type for "program accounts" queries from that of the return type of `getProgramAccounts`. This is an intentional and conscious design choice made in order to fully align query-ability and expected payloads with the types of the accounts schema.

TL/DR, I stuck `pubkey` inside the actual account payload itself as `address` because I think it makes more sense, especially considering the `Account` GraphQL interface already has this value.