refactor(experimental): store fee payer Signers in transactions (#2111)
### Context

The `@solana/signers` package allows us to store `TransactionSigners` inside instructions' account meta which can then be extracted in order to sign the transaction they belong to.

### Problem

Signing transactions that way currently doesn't take the fee payer into account. It works most of the time as it is likely that the fee payer will end up being required as a signer inside one of the instructions. However, when that's not the case, the fee payer will have to manually sign the transaction on top of using the `signTransactionWithSigners` function.

### Solution

This PR fixes that by storing the fee payer signer inside the transaction object, much like we do with the instruction account metas.

It offers a new `setTransactionFeePayerSigner` which, on top of adding the `feePayer` address, also add a `feePayerSigner` attribute that stores the fee payer as a `TransactionSigner`.

That way, when using the `@solana/signers` package, you get to attach all relevant signers to a transaction and then use `signTransactionWithSigners` to handle the whole signing process, including the fee payer.