Replace `RpcError` with a coded exception (#2286)
# Summary

So wow. I completely missed these as part of #2118, and it turns out they're a really big deal and required a ton of changes.

There are:

* Errors that have enough data to format a message
* Errors that have no data, but also no context in the message
* Errors that have no data, but really should, because you can't format a message without it
* Preflight errors in which is nested a `TransactionError` (see #2213)

In this PR we create a helper that takes in the `RpcSimulateTransactionResult` from the RPC and reformats it as a coded `SolanaError`.

As always, everything you need to know is in the `packages/errors/src/__tests__/json-rpc-error-test.ts`.

# Test Plan

```
pnpm turbo test:unit:browser
pnpm turbo test:unit:node
```