refactor(experimental): add lamports codec
Similar to the commit before it, this change adds a codec to `@solana/rpc-types`
for working with the `LamportsUnsafeBeyond2Pow53Minus1` opaque type.

It's not enough to just use the `u64Codec` directly, since TypeScript won't just
infer `LamportsUnsafeBeyond2Pow53Minus1` from a `bigint`. So we have to set up a
`mapCodec` for this coercion.